+++ packages/contracts/src/core/plugin/dao-authorizable/DaoAuthorizableUpgradeable.sol
@@ -9,29 +9,29 @@ import {_auth} from "../../utils/auth.sol";
 
 /// @title DaoAuthorizableUpgradeable
 /// @author Aragon Association - 2022-2023
-/// @notice An abstract contract providing a meta-transaction compatible modifier for upgradeable contracts to authorize function calls through an associated DAO.
+/// @notice An abstract contract providing a meta-transaction compatible modifier for upgradeable or cloneable contracts to authorize function calls through an associated DAO.
 /// @dev Make sure to call `__DaoAuthorizableUpgradeable_init` during initialization of the inheriting contract.
 abstract contract DaoAuthorizableUpgradeable is ContextUpgradeable {
     /// @notice The associated DAO managing the permissions of inheriting contracts.
-    IDAO internal dao;
+    IDAO private dao_;
+
+    /// @notice Initializes the contract by setting the associated DAO.
+    /// @param _dao The associated DAO address.
+    function __DaoAuthorizableUpgradeable_init(IDAO _dao) internal onlyInitializing {
+        dao_ = _dao;
+    }
 
     /// @notice Returns the DAO contract.
     /// @return IDAO The DAO contract.
-    function getDAO() external view returns (IDAO) {
-        return dao;
+    function dao() public view returns (IDAO) {
+        return dao_;
     }
 
     /// @notice A modifier to be used to check permissions on a target contract via the associated DAO.
     /// @param _permissionId The permission identifier required to call the method this modifier is applied to.
     modifier auth(bytes32 _permissionId) {
-        _auth(dao, address(this), _msgSender(), _permissionId, _msgData());
+        _auth(dao_, address(this), _msgSender(), _permissionId, _msgData());
         _;
-    }
-
-    /// @notice Initializes the contract by setting the associated DAO.
-    /// @param _dao The associated DAO address.
-    function __DaoAuthorizableUpgradeable_init(IDAO _dao) internal onlyInitializing {
-        dao = _dao;
     }
 
     /// @notice This empty reserved space is put in place to allow future versions to add new variables without shifting down storage in the inheritance chain (see [OpenZepplins guide about storage gaps](https://docs.openzeppelin.com/contracts/4.x/upgradeable#storage_gaps)).